+++ client/src/app/bounties/[id]/page.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Layout from '@/components/layout/Layout';
 import { Bounty, Difficulty, BountyStatus, BountyCategory } from '@/types';
 import { Card, CardContent } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
