+++ client/src/app/bounties/[id]/submit/page.tsx
@@ -1,6 +1,5 @@
 'use client';
 import React, { useState } from 'react';
-import Layout from '@/components/layout/Layout';
 import { Card, CardContent } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
