+++ client/src/app/bounties/dashboard/page.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Layout from '@/components/layout/Layout';
 import { Card, CardContent } from '@/components/ui/card';
 import { Trophy, DollarSign, TrendingUp } from 'lucide-react';
